perm check in api

+++ administrator/components/com_media/resources/scripts/store/actions.es6.js
@@ -129,6 +129,9 @@ export const toggleBrowserItemSelect = (context, payload) => {
  * @param payload object with the new folder name and its parent directory
  */
 export const createDirectory = (context, payload) => {
+  if (!api.canCreate) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   api.createDirectory(payload.name, payload.parent)
     .then((folder) => {
@@ -150,6 +153,9 @@ export const createDirectory = (context, payload) => {
  * @param payload object with the new folder name and its parent directory
  */
 export const uploadFile = (context, payload) => {
+  if (!api.canEdit) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   api.upload(payload.name, payload.parent, payload.content, payload.override || false)
     .then((file) => {
@@ -175,6 +181,9 @@ export const uploadFile = (context, payload) => {
  * @param payload object: the old and the new path
  */
 export const renameItem = (context, payload) => {
+  if (!api.canEdit) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   api.rename(payload.path, payload.newPath)
     .then((item) => {
@@ -199,11 +208,17 @@ export const renameItem = (context, payload) => {
  * @param context
  */
 export const deleteSelectedItems = (context) => {
+  if (!api.canDelete) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   // Get the selected items from the store
   const { selectedItems } = context.state;
   if (selectedItems.length > 0) {
     selectedItems.forEach((item) => {
+      if (typeof item.canDelete !== 'undefined' && item.canDelete === false) {
+        return;
+      }
       api.delete(item.path)
         .then(() => {
           context.commit(types.DELETE_SUCCESS, item);
